+++ src/hooks.js
@@ -1,5 +1,6 @@
 import { useState, useMemo } from 'react'
 import {
+  setInputValue,
   updateInputValue,
   resetInputValue,
   clearInputErrors,
@@ -12,13 +13,14 @@ import {
  * @param {Object} initialFormValues
  * Where each key is the form input name,
  * and the value is the initial value of the form input.
+ * @example {name: "initialValue"}
  * @returns
  * - formValues - Object where each key is the input name and each value is the input value.
  * - resetFormValues - Function that resets the form values to the initial state.
- * - resetInputValue - Function that resets the input value to it's initial state.
- * - updateInputValue - Function called via an event handler that updates an input value.
- * - setInputValue - Function that updates an input value.
- * - setFormValues - Function that updates the form values.
+ * - resetInputValue(name) - Function that resets the input value to it's initial state.
+ * - updateInputValue - (deprecated) please use `setInputValue` instead.
+ * - setInputValue(input, value?) - Function that updates an input value via an event handler or passing the new value.
+ * - setFormValues(formValues) - Function that updates the form values.
  */
 export function useFormValues(initialFormValues) {
   const [formValues, setFormValues] = useState(initialFormValues)
@@ -31,12 +33,15 @@ export function useFormValues(initialFormValues) {
      */
     resetInputValue: name =>
       resetInputValue(name, setFormValues, initialFormValues),
+    /**
+     * @deprecated - please use `setInputValue`
+     */
     updateInputValue: event => updateInputValue(event, setFormValues),
     /**
-     * @param {String} name
+     * @param {String|import('react').SyntheticEvent} input
+     * @param {*=} value
      */
-    setInputValue: (name, value) =>
-      setFormValues(prevFormValues => ({ ...prevFormValues, [name]: value })),
+    setInputValue: (input, value) => setInputValue(input, value, setFormValues),
     setFormValues
   }
 }
@@ -50,15 +55,15 @@ export function useFormValues(initialFormValues) {
  * - formErrors - Object where each key is the input name and each value is an array of error messages (string/jsx).
  * - numberOfErrors - Number of inputs that have errors.
  * - validateForm - Function that runs validation on the whole form, returns formIsValid (boolean).
- * - validateInputValue - Function that runs validation on an input.
+ * - validateInputValue(input) - Function that runs validation on an input.
  * - clearFormErrors - Function that clears the forms errors.
- * - clearInputErrors - Function that clears the errors for an input.
- * - setInputErrors - Function that sets the errors for an input.
+ * - clearInputErrors(input) - Function that clears the errors for an input via an event handler or passing the new value.
+ * - setInputErrors(name, errors) - Function that sets the errors for an input.
  */
 export function useFormErrors(formValidations = {}) {
   const initialFormErrors = useMemo(
     () => getInitialFormErrors(formValidations),
-    []
+    [formValidations]
   )
   const [formErrors, setFormErrors] = useState(initialFormErrors)
 
@@ -68,13 +73,17 @@ export function useFormErrors(formValidations = {}) {
       .length,
     validateForm: formValues =>
       validateForm(formValues, setFormErrors, formValidations),
-    validateInputValue: event =>
-      validateInputValue(event, setFormErrors, formValidations),
+    /**
+     * @param {String|import('react').SyntheticEvent} input
+     * @param {*=} value
+     */
+    validateInputValue: (input, value = null) =>
+      validateInputValue(input, value, setFormErrors, formValidations),
     clearFormErrors: () => setFormErrors(initialFormErrors),
     /**
-     * @param {String} name
+     * @param {String|import('react').SyntheticEvent} input
      */
-    clearInputErrors: name => clearInputErrors(name, setFormErrors),
+    clearInputErrors: input => clearInputErrors(input, setFormErrors),
     /**
      * @param {String} name
      * @param {String[]} errors